style: Make pages folder and Move page components in it
